refactor: magic-number => display-width 파일명 변경
