+++ client/src/utils/display-width.ts
